pyproject.toml
[project]
name = "initial-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "google-genai>=1.22.0",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "replicate>=1.0.7",
    "requests>=2.32.4",
]

[dependency-groups]
dev = ["pytest>=8.4.0"]


[tool.pytest.ini_options]
pythonpath = ["src", "scripts"]

[tool.setuptools]
package-dir = { "" = "src" }
